requirements.txt
django==1.11.6
mysqlclient==1.3.12
ipdb==0.10.3
ipython==4.1.2
django-extensions==1.9.6
django-jsonview==1.1.0
coverage==4.4.1
